refactor: Line 서비스에 있는 Section 로직을 Section 서비스로 이동
